[Refactor] #963 Moved calculate_upkeep() from player-status.h to pet-util.h
