+++ src/effect/effect-monster-charm.cpp
@@ -17,8 +17,8 @@
 #include "monster/monster-status.h"
 #include "object-enchant/trc-types.h"
 #include "pet/pet-fall-off.h"
+#include "pet/pet-util.h"
 #include "player-info/avatar.h"
-#include "player/player-status.h"
 #include "player/player-status-flags.h"
 #include "util/bit-flags-calculator.h"
 #include "spell/spells-diceroll.h"
