+++ src/knowledge/knowledge-monsters.cpp
@@ -25,7 +25,7 @@
 #include "monster/monster-info.h"
 #include "monster/monster-status.h"
 #include "monster/smart-learn-types.h"
-#include "player/player-status.h"
+#include "pet/pet-util.h"
 #include "system/floor-type-definition.h"
 #include "system/monster-race-definition.h"
 #include "system/monster-type-definition.h"
